# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MyMovies Extractor API",
    "version": "1.0.0",
    "description": "REST API per estrazione recensioni da MyMovies.it"
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local" }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HealthResponse" }
              }
            }
          }
        }
      }
    },
    "/api/info": {
      "get": {
        "summary": "Informazioni API",
        "responses": {
          "200": {
            "description": "Informazioni API",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InfoResponse" }
              }
            }
          }
        }
      }
    },
    "/api/extract": {
      "post": {
        "summary": "Estrai una singola recensione",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ExtractRequest" },
              "examples": {
                "default": {
                  "value": { "title": "Oppenheimer", "year": 2023, "options": { "noSave": false } }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Recensione estratta",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ExtractSuccessResponse" } }
            }
          },
          "400": {
            "description": "Richiesta non valida",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "Recensione non trovata",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ExtractNotFoundResponse" } } }
          },
          "429": {
            "description": "Too Many Requests",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RateLimitError" } } }
          },
          "500": {
            "description": "Errore interno",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/api/extract/batch": {
      "post": {
        "summary": "Estrazione multipla (massimo 10 film)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BatchRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Risultati batch",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BatchResponse" }
              }
            }
          },
          "400": {
            "description": "Richiesta non valida",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "429": {
            "description": "Too Many Requests",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RateLimitError" } } }
          },
          "500": {
            "description": "Errore interno",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/api/reviews": {
      "get": {
        "summary": "Lista recensioni estratte",
        "responses": {
          "200": {
            "description": "Lista",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ReviewsListResponse" }
              }
            }
          },
          "500": {
            "description": "Errore interno",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/api/reviews/{filename}": {
      "get": {
        "summary": "Recensione specifica (testo o JSON)",
        "parameters": [
          { "name": "filename", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Nome file .txt" },
          { "name": "format", "in": "query", "required": false, "schema": { "type": "string", "enum": ["text", "json"] }, "description": "Formato output" }
        ],
        "responses": {
          "200": {
            "description": "Contenuto recensione",
            "content": {
              "text/plain": { "schema": { "type": "string" } },
              "application/json": { "schema": { "$ref": "#/components/schemas/ReviewFileJsonResponse" } }
            }
          },
          "400": {
            "description": "Nome file non valido",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "File non trovato",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "500": {
            "description": "Errore interno",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/api/stats": {
      "get": {
        "summary": "Statistiche recensioni",
        "responses": {
          "200": {
            "description": "Statistiche",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/StatsResponse" } }
            }
          },
          "500": {
            "description": "Errore interno",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/api/openapi.json": {
      "get": {
        "summary": "Specifica OpenAPI",
        "responses": {
          "200": {
            "description": "Spec JSON",
            "content": { "application/json": { "schema": { "type": "object" } } }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HealthResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "example": "ok" },
          "service": { "type": "string", "example": "MyMovies API" },
          "version": { "type": "string", "example": "1.0.0" },
          "timestamp": { "type": "string", "format": "date-time" }
        },
        "required": ["status", "service", "version", "timestamp"]
      },
      "InfoResponse": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "version": { "type": "string" },
          "description": { "type": "string" },
          "endpoints": { "type": "object", "additionalProperties": { "type": "string" } },
          "rate_limit": { "type": "string" },
          "documentation": { "type": "string" }
        }
      },
      "ExtractRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "year": { "type": "integer", "minimum": 1900, "maximum": 2030 },
          "options": {
            "type": "object",
            "properties": {
              "noSave": { "type": "boolean", "default": false },
              "headless": { "type": "boolean", "default": true }
            },
            "additionalProperties": true
          }
        },
        "required": ["title", "year"]
      },
      "ExtractSuccessResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "const": true },
          "data": {
            "type": "object",
            "properties": {
              "title": { "type": "string" },
              "year": { "type": "integer" },
              "review": {
                "type": "object",
                "properties": {
                  "content": { "type": "string" },
                  "author": { "type": ["string", "null"] },
                  "date": { "type": ["string", "null"] }
                },
                "required": ["content"]
              },
              "metadata": {
                "type": "object",
                "properties": {
                  "extractionMethod": { "type": ["string", "null"] },
                  "contentLength": { "type": "integer" },
                  "wordCount": { "type": "integer" },
                  "processingTime": { "type": "integer" },
                  "apiVersion": { "type": ["string", "null"] }
                }
              },
              "url": { "type": ["string", "null"], "format": "uri" },
              "filePath": { "type": ["string", "null"] }
            },
            "required": ["title", "year", "review", "metadata"]
          }
        },
        "required": ["success", "data"]
      },
      "ExtractNotFoundResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "const": false },
          "error": { "type": "string", "example": "Not Found" },
          "message": { "type": "string" },
          "url": { "type": ["string", "null"], "format": "uri" }
        },
        "required": ["success", "error", "message"]
      },
      "BatchRequest": {
        "type": "object",
        "properties": {
          "films": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "title": { "type": "string" },
                "year": { "type": "integer", "minimum": 1900, "maximum": 2030 }
              },
              "required": ["title", "year"]
            },
            "minItems": 1,
            "maxItems": 10
          },
          "options": { "$ref": "#/components/schemas/ExtractRequest/properties/options" }
        },
        "required": ["films"]
      },
      "BatchResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "batch": {
            "type": "object",
            "properties": {
              "total": { "type": "integer" },
              "successful": { "type": "integer" },
              "failed": { "type": "integer" },
              "processingTime": { "type": "integer" }
            }
          },
          "results": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "index": { "type": "integer" },
                "success": { "type": "boolean" },
                "data": {
                  "type": ["object", "null"],
                  "properties": {
                    "title": { "type": "string" },
                    "year": { "type": "integer" },
                    "author": { "type": ["string", "null"] },
                    "contentLength": { "type": "integer" },
                    "filePath": { "type": ["string", "null"] }
                  }
                },
                "error": { "type": ["string", "null"] },
                "input": {
                  "type": "object",
                  "properties": { "title": { "type": "string" }, "year": { "type": "integer" } }
                }
              }
            }
          }
        }
      },
      "ReviewsListResponse": {
        "type": "object",
        "properties": {
          "reviews": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ReviewInfo" }
          },
          "total": { "type": "integer" }
        }
      },
      "ReviewInfo": {
        "type": "object",
        "properties": {
          "filename": { "type": "string" },
          "title": { "type": "string" },
          "year": { "type": ["integer", "null"] },
          "size": { "type": "integer" },
          "modified": { "type": "string", "format": "date-time" },
          "downloadUrl": { "type": "string" }
        }
      },
      "ReviewFileJsonResponse": {
        "type": "object",
        "properties": {
          "filename": { "type": "string" },
          "title": { "type": "string" },
          "author": { "type": "string" },
          "date": { "type": "string" },
          "content": { "type": "string" },
          "fullText": { "type": "string" }
        }
      },
      "StatsResponse": {
        "type": "object",
        "properties": {
          "totalFiles": { "type": "integer" },
          "totalSize": { "type": "integer" },
          "averageSize": { "type": "integer" },
          "oldestReview": { "type": ["string", "null"] },
          "newestReview": { "type": ["string", "null"] },
          "timestamp": { "type": "string", "format": "date-time" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "const": false },
          "error": { "type": "string" },
          "message": { "type": "string" }
        }
      },
      "RateLimitError": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "Too many requests" },
          "message": { "type": "string" }
        }
      }
    }
  }
}